Remove logic to bind OrderByClause from the FilterBinder to a new OrderByBinder
