+++ src/Microsoft.AspNetCore.OData/Query/Expressions/IOrderByBinder.cs
@@ -1,5 +1,5 @@
 ﻿//-----------------------------------------------------------------------------
-// <copyright file="IFilterBinder.cs" company=".NET Foundation">
+// <copyright file="IOrderByBinder.cs" company=".NET Foundation">
 //      Copyright (c) .NET Foundation and Contributors. All rights reserved.
 //      See License.txt in the project root for license information.
 // </copyright>
@@ -15,22 +15,14 @@ namespace Microsoft.AspNetCore.OData.Query.Expressions
     /// Exposes the ability to translate an OData $filter parse tree represented by <see cref="FilterClause"/> to
     /// an <see cref="Expression"/> and applies it to an <see cref="IQueryable"/>.
     /// </summary>
-    public interface IFilterBinder
+    public interface IOrderByBinder
     {
         /// <summary>
-        /// Applies an OData $filter parse tree represented by <see cref="FilterClause"/> to an <see cref="IQueryable"/>.
+        /// Applies an OData $orderby parse tree represented by <see cref="OrderByClause"/> to an <see cref="IQueryable"/>.
         /// </summary>
         /// <param name="source">The original <see cref="IQueryable"/>.</param>
-        /// <param name="context">An instance of the <see cref="FilterBinderContext"/>.</param>
+        /// <param name="context">An instance of the <see cref="OrderByBinderContext"/>.</param>
         /// <returns>The new <see cref="IQueryable"/> after the filter query has been applied to.</returns>
-        IQueryable Bind(IQueryable source, FilterBinderContext context);
-
-        /// <summary>
-        /// Creates an <see cref="Expression"/> from an OData $filter parse tree represented by <see cref="FilterClause"/>.
-        /// </summary>
-        /// <param name="source">The original <see cref="IQueryable"/>.</param>
-        /// <param name="context">An instance of the <see cref="FilterBinderContext"/>.</param>
-        /// <returns>An <see cref="Expression"/> which can be later applied to an <see cref="IQueryable"/>.</returns>
-        Expression BindFilterClause(IQueryable source, FilterBinderContext context);
+        Expression Bind(IQueryable source, OrderByBinderContext context);
     }
 }
