+++ src/Microsoft.AspNetCore.OData/Query/Expressions/OrderByBinderContext.cs
@@ -1,5 +1,5 @@
 ﻿//-----------------------------------------------------------------------------
-// <copyright file="FilterBinderContext.cs" company=".NET Foundation">
+// <copyright file="OrderByBinderContext.cs" company=".NET Foundation">
 //      Copyright (c) .NET Foundation and Contributors. All rights reserved.
 //      See License.txt in the project root for license information.
 // </copyright>
@@ -12,14 +12,14 @@ using Microsoft.OData.UriParser;
 namespace Microsoft.AspNetCore.OData.Query.Expressions
 {
     /// <summary>
-    /// Wrapper for properties used by the <see cref="IFilterBinder"/>.
+    /// Wrapper for properties used by the <see cref="IOrderByBinder"/>.
     /// </summary>
-    public class FilterBinderContext
+    public class OrderByBinderContext
     {
         /// <summary>
-        /// Gets or sets the parsed <see cref="Microsoft.OData.UriParser.FilterClause"/> for this query option.
+        /// Gets or sets the parsed <see cref="Microsoft.OData.UriParser.OrderByClause"/> for this query option.
         /// </summary>
-        public FilterClause FilterClause { get; set; }
+        public OrderByClause OrderByClause { get; set; }
 
         /// <summary>
         /// Gets or sets the <see cref="ODataQuerySettings"/> that contains all the query application related settings.
