+++ test/Microsoft.AspNetCore.OData.Tests/PublicApi/Microsoft.AspNetCore.OData.PublicApi.NetCore31.bsl
@@ -2448,7 +2448,10 @@ public class Microsoft.AspNetCore.OData.Query.Container.TruncatedCollection`1 : 
 public interface Microsoft.AspNetCore.OData.Query.Expressions.IFilterBinder {
 	System.Linq.IQueryable Bind (System.Linq.IQueryable source, Microsoft.AspNetCore.OData.Query.Expressions.FilterBinderContext context)
 	System.Linq.Expressions.Expression BindFilterClause (System.Linq.IQueryable source, Microsoft.AspNetCore.OData.Query.Expressions.FilterBinderContext context)
-	System.Linq.Expressions.Expression BindOrderByClause (System.Linq.IQueryable source, Microsoft.AspNetCore.OData.Query.Expressions.FilterBinderContext context)
+}
+
+public interface Microsoft.AspNetCore.OData.Query.Expressions.IOrderByBinder {
+	System.Linq.Expressions.Expression Bind (System.Linq.IQueryable source, Microsoft.AspNetCore.OData.Query.Expressions.OrderByBinderContext context)
 }
 
 public interface Microsoft.AspNetCore.OData.Query.Expressions.ISelectExpandBinder {
@@ -2471,13 +2474,26 @@ public abstract class Microsoft.AspNetCore.OData.Query.Expressions.ExpressionBin
 	protected System.Linq.Expressions.Expression GetFlattenedPropertyExpression (string propertyPath)
 }
 
-public class Microsoft.AspNetCore.OData.Query.Expressions.FilterBinder : Microsoft.AspNetCore.OData.Query.Expressions.ExpressionBinderBase, IFilterBinder {
+public class Microsoft.AspNetCore.OData.Query.Expressions.FilterBinder : Microsoft.AspNetCore.OData.Query.Expressions.FilterOrderByBinderBase, IFilterBinder {
 	public FilterBinder (System.IServiceProvider requestContainer)
 
+	public virtual System.Linq.IQueryable Bind (System.Linq.IQueryable source, Microsoft.AspNetCore.OData.Query.Expressions.FilterBinderContext context)
+	public virtual System.Linq.Expressions.Expression BindFilterClause (System.Linq.IQueryable source, Microsoft.AspNetCore.OData.Query.Expressions.FilterBinderContext context)
+}
+
+public class Microsoft.AspNetCore.OData.Query.Expressions.FilterBinderContext {
+	public FilterBinderContext ()
+
+	System.Type ElementClrType  { public get; public set; }
+	Microsoft.OData.UriParser.FilterClause FilterClause  { public get; public set; }
+	Microsoft.AspNetCore.OData.Query.ODataQueryContext QueryContext  { public get; public set; }
+	Microsoft.AspNetCore.OData.Query.ODataQuerySettings QuerySettings  { public get; public set; }
+}
+
+public class Microsoft.AspNetCore.OData.Query.Expressions.FilterOrderByBinderBase : Microsoft.AspNetCore.OData.Query.Expressions.ExpressionBinderBase {
 	System.Linq.Expressions.ParameterExpression Parameter  { protected virtual get; }
 
 	public virtual System.Linq.Expressions.Expression Bind (Microsoft.OData.UriParser.QueryNode node)
-	public virtual System.Linq.IQueryable Bind (System.Linq.IQueryable source, Microsoft.AspNetCore.OData.Query.Expressions.FilterBinderContext context)
 	public virtual System.Linq.Expressions.Expression BindAllNode (Microsoft.OData.UriParser.AllNode allNode)
 	public virtual System.Linq.Expressions.Expression BindAnyNode (Microsoft.OData.UriParser.AnyNode anyNode)
 	public virtual System.Linq.Expressions.Expression BindBinaryOperatorNode (Microsoft.OData.UriParser.BinaryOperatorNode binaryOperatorNode)
@@ -2487,11 +2503,9 @@ public class Microsoft.AspNetCore.OData.Query.Expressions.FilterBinder : Microso
 	public virtual System.Linq.Expressions.Expression BindConvertNode (Microsoft.OData.UriParser.ConvertNode convertNode)
 	public virtual System.Linq.Expressions.Expression BindCountNode (Microsoft.OData.UriParser.CountNode node)
 	public virtual System.Linq.Expressions.Expression BindDynamicPropertyAccessQueryNode (Microsoft.OData.UriParser.SingleValueOpenPropertyAccessNode openNode)
-	public virtual System.Linq.Expressions.Expression BindFilterClause (System.Linq.IQueryable source, Microsoft.AspNetCore.OData.Query.Expressions.FilterBinderContext context)
 	public virtual System.Linq.Expressions.Expression BindInNode (Microsoft.OData.UriParser.InNode inNode)
 	public virtual System.Linq.Expressions.Expression BindNavigationPropertyNode (Microsoft.OData.UriParser.QueryNode sourceNode, Microsoft.OData.Edm.IEdmNavigationProperty navigationProperty)
 	public virtual System.Linq.Expressions.Expression BindNavigationPropertyNode (Microsoft.OData.UriParser.QueryNode sourceNode, Microsoft.OData.Edm.IEdmNavigationProperty navigationProperty, string propertyPath)
-	public virtual System.Linq.Expressions.Expression BindOrderByClause (System.Linq.IQueryable source, Microsoft.AspNetCore.OData.Query.Expressions.FilterBinderContext context)
 	public virtual System.Linq.Expressions.Expression BindPropertyAccessQueryNode (Microsoft.OData.UriParser.SingleValuePropertyAccessNode propertyAccessNode)
 	public virtual System.Linq.Expressions.Expression BindRangeVariable (Microsoft.OData.UriParser.RangeVariable rangeVariable)
 	public virtual System.Linq.Expressions.Expression BindSingleComplexNode (Microsoft.OData.UriParser.SingleComplexNode singleComplexNode)
@@ -2501,11 +2515,14 @@ public class Microsoft.AspNetCore.OData.Query.Expressions.FilterBinder : Microso
 	public virtual System.Linq.Expressions.Expression BindUnaryOperatorNode (Microsoft.OData.UriParser.UnaryOperatorNode unaryOperatorNode)
 }
 
-public class Microsoft.AspNetCore.OData.Query.Expressions.FilterBinderContext {
-	public FilterBinderContext ()
+public class Microsoft.AspNetCore.OData.Query.Expressions.OrderByBinder : Microsoft.AspNetCore.OData.Query.Expressions.FilterOrderByBinderBase, IOrderByBinder {
+	public virtual System.Linq.Expressions.Expression Bind (System.Linq.IQueryable source, Microsoft.AspNetCore.OData.Query.Expressions.OrderByBinderContext context)
+}
+
+public class Microsoft.AspNetCore.OData.Query.Expressions.OrderByBinderContext {
+	public OrderByBinderContext ()
 
 	System.Type ElementClrType  { public get; public set; }
-	Microsoft.OData.UriParser.FilterClause FilterClause  { public get; public set; }
 	Microsoft.OData.UriParser.OrderByClause OrderByClause  { public get; public set; }
 	Microsoft.AspNetCore.OData.Query.ODataQueryContext QueryContext  { public get; public set; }
 	Microsoft.AspNetCore.OData.Query.ODataQuerySettings QuerySettings  { public get; public set; }
